notification: Don't inhibit notifs with the same summary after close
If SingleNotification::close() is called, a subsequent call to show()
with the same summary will fail to display the notification, because the
summary field is not reset, causing the comparison in `show()` to block
re-creation.

This can prevent (for example) low battery notifications from displaying
again after being closed, as described in #79.

The summary field is now an Option<String> and is cleared in close(),
allowing notifications with the same summary to be displayed after the
previous one is closed.

Fixes #79.